Addresses Issue #76. Normalized "Reply" with JS codebase.
