+++ libraries/Microsoft.Bot.Builder/IBotContext.cs
@@ -51,11 +51,19 @@ namespace Microsoft.Bot.Builder
         bool IfIntent(Regex expression);
 
         /// <summary>
-        /// Send a reply to the sender
+        /// Queues a new "message" responses array.
         /// </summary>
-        /// <param name="text"></param>
+        /// <param name="text">Text of a message to send to the user.</param>
+        /// <param name="speak">(Optional) SSML that should be spoken to the user on channels that support speech.</param>
         /// <returns></returns>
-        BotContext Reply(string text);
+        BotContext Reply(string text, string speak = null);
+
+        /// <summary>
+        /// Queues a new "message" responses array.
+        /// </summary>
+        /// <param name="activity">Activity object to send to the user.</param>
+        /// <returns></returns>
+        BotContext Reply(IActivity activity);
 
         /// <summary>
         /// Send a reply using a templateId bound to data
@@ -78,107 +86,6 @@ namespace Microsoft.Bot.Builder
         public static BotContext ToBotContext(this IBotContext context)
         {
             return (BotContext)context; 
-        }
-    }
-
-    public class BotContext : FlexObject, IBotContext
-    {
-        private readonly Bot _bot;
-        private readonly IActivity _request;        
-        private readonly ConversationReference _conversationReference;
-        private readonly BotState _state = new BotState();
-        private IList<IActivity> _responses = new List<IActivity>();
-
-        public BotContext(Bot bot, IActivity request)
-        {
-            _bot = bot ?? throw new ArgumentNullException(nameof(bot));
-            _request = request ?? throw new ArgumentNullException(nameof(request)); 
-
-            _conversationReference = new ConversationReference()
-            {
-                ActivityId = request.Id,
-                User = request.From,
-                Bot = request.Recipient,
-                Conversation = request.Conversation,
-                ChannelId = request.ChannelId,
-                ServiceUrl = request.ServiceUrl
-            };
-        }
-
-        public BotContext(Bot bot, ConversationReference conversationReference)
-        {
-            _bot = bot ?? throw new ArgumentNullException(nameof(bot));
-            _conversationReference = conversationReference ?? throw new ArgumentNullException(nameof(conversationReference));
-        }
-
-        public async Task PostActivity(IBotContext context, IList<IActivity> activities)
-        {
-            await _bot.PostActivity(context, activities).ConfigureAwait(false);            
-        }
-
-        public IActivity Request => _request;
-
-        public Bot Bot => _bot;
-
-        public IList<IActivity> Responses { get => _responses; set => this._responses = value; }
-
-        public IStorage Storage { get; set; }
-
-        public Intent TopIntent { get; set; }
-
-        public TemplateManager TemplateManager { get; set; }
-
-        public bool IfIntent(string intentName)
-        {
-            if (string.IsNullOrWhiteSpace(intentName))
-                throw new ArgumentNullException(nameof(intentName)); 
-
-            if (this.TopIntent != null)
-            {
-                if (TopIntent.Name == intentName)
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-        public bool IfIntent(Regex expression)
-        {
-            if (expression == null)
-                throw new ArgumentNullException(nameof(expression)); 
-
-            if (this.TopIntent != null)
-            {
-                if (expression.IsMatch(this.TopIntent.Name))
-                    return true;
-            }
-
-            return false;
-        }
-
-
-        public ConversationReference ConversationReference { get => _conversationReference; }
-
-        public BotState State { get => _state; }
-
-        public BotContext Reply(string text)
-        {
-            var reply = this.ConversationReference.GetPostToUserMessage();
-            reply.Text = text;
-            this.Responses.Add(reply);
-            return this; 
-        }
-
-        public BotContext ReplyWith(string templateId, object data=null)
-        {
-            // queue template activity to be databound when sent
-            var reply = this.ConversationReference.GetPostToUserMessage();
-            reply.Type = TemplateManager.TEMPLATE;
-            reply.Text = templateId;
-            reply.Value = data;
-            this.Responses.Add(reply);
-            return this;
         }
     }
 }
